Fix build error.
Upstream API has recently changed vcsclient.FileWithRange.StartByte
type from int to int64, hence the casts are needed now.
Fixes Travis build error.

+++ highlight_test.go
@@ -159,7 +159,7 @@ func TestCodeEquals(t *testing.T) {
 func newFileWithRange(src []byte) *vcsclient.FileWithRange {
 	return &vcsclient.FileWithRange{
 		TreeEntry: &vcsclient.TreeEntry{Contents: []byte(src)},
-		FileRange: vcsclient.FileRange{StartByte: 0, EndByte: len(src)},
+		FileRange: vcsclient.FileRange{StartByte: 0, EndByte: int64(len(src))},
 	}
 }
